Overhaul training but changing the inputs
Changed the inputs to be more around decision making instead of simply the raw board state. The idea is that it'll be more easier for the network to get to a decent proficency faster than waiting for it build a complex network off of the raw board state. Also fixed a few bugs here and there