Improve fitness algorithm
Before this change, games that ended because all but one player had a negative fitness from not making moves, the player with the highest fitness and their partner would receive a win. Though this likely didn't affect much, it makes more sense to only award the win to the player with a positive fitness. In the event the game ends where both partners finish with a postive fitness then both receive a win. This was added for when games do actually make it to the end. Additionally pre-seeded the winner map with all players so that if only one player receives a win during the session then two will still advanced in order to keep the bracket properly balanced.