Increase number of round robins and population size
After a while it doesn't seem like it's evolving. This could be due to a better mutation not happening or there being too much variation to recognize an improvement.
Increasing the round robin amount to try to reduce variation and increasing population size to increase the chances of a successful mutation