Update staleness logic
Previous staleness logic is very punishing since the only way to reset it is to get better quickly. In order to allow the AI to evolve properly while also mitigating staleness the logic has been changed so that a species has to not be last 5 straight times to reset the staleness counter. If it's unable to do that before acruing 50 last place finishes then it will be removed from the pool