Rebuild training algorithm
Instead of doing a bracket where a top genome could fluke out and be removed from the pool, this now has it so that each genome in a species playes against the top genome of the species. At the end of that they're sorted from least to most wins where the previous top genome is placed between the genome that had the most wins but lost to the top genome and the genome that ahd the least wins but beat the top genome. After that the new top genomes of each top species play round robin. Additionally each matchup is one on one which redues variability. This increases the total number of games needed to be played by a slight amount but, in return, makes it very incredibly unlikely that a "better" genome is removed from the pool over a "worse" genome.